Refactor file handling and notification services
- Removed `IFileService` and introduced `IImportFileService` and `IExportFileService` for better separation of concerns.
- Updated `IMedicalEventNotificationService` to include `SendMedicalEventNotificationToManagerAsync`.
- Changed return type of `CreateMedicalEventAsync` in `IMedicalEventService` to `NotificationMedicalEventResponse`.
- Replaced `FileController` with `ImportFileController` and `ExportFileController` for handling file operations.
- Split `FileService` into `ImportFileService` and `ExportFileService`.
- Enhanced `MedicalEventNotificationService` to send notifications to both managers and parents.
- Added new methods for exporting student, medical inventory, and vaccination details to Excel files.
- Introduced `NotificationMedicalEventResponse` class for encapsulating notification requests.
