+++ SchoolMedicalServer.Abstractions/Dtos/Notification/NotificationMedicalEventResponse.cs
@@ -1,0 +1,8 @@
+﻿namespace SchoolMedicalServer.Abstractions.Dtos.Notification
+{
+    public class NotificationMedicalEventResponse
+    {
+        public NotificationRequest? ToParent { get; set; }
+        public NotificationRequest? ToManager { get; set; }
+    }
+}
