+++ SchoolMedicalServer.Abstractions/IServices/IExportFileService.cs
@@ -1,0 +1,9 @@
+﻿namespace SchoolMedicalServer.Abstractions.IServices
+{
+    public interface IExportFileService
+    {
+        Task<byte[]> ExportStudentsExcelFileAsync();
+        Task<byte[]> ExportMedicalInventoriesExcelFileAsync();
+        Task<byte[]> ExportVaccinationDetailsExcelFileAsync();
+    }
+}
