+++ SchoolMedicalServer.Abstractions/IServices/IMedicalEventNotificationService.cs
@@ -4,6 +4,7 @@ namespace SchoolMedicalServer.Abstractions.IServices
 {
     public interface IMedicalEventNotificationService
     {
+        Task<NotificationResponse> SendMedicalEventNotificationToManagerAsync(NotificationRequest request);
         Task<NotificationResponse> SendMedicalEventNotificationToParentAsync(NotificationRequest request);
     }
 }
