+++ SchoolMedicalServer.Infrastructure/Services/ImportFileService.cs
@@ -6,12 +6,12 @@ using SchoolMedicalServer.Abstractions.IServices;
 
 namespace SchoolMedicalServer.Infrastructure.Services
 {
-    public class FileService(
+    public class ImportFileService(
         IStudentRepository studentRepository,
         IMedicalInventoryRepository medicalInventoryRepository,
         IHealthProfileRepository healthProfileRepository,
         IBaseRepository baseRepository,
-        IVaccinationDetailsRepository vacctionDetailsRepository) : IFileService
+        IVaccinationDetailsRepository vaccinationDetailsRepository) : IImportFileService
     {
         public async Task UploadMedicalInventoriesExcelFile(IFormFile file)
         {
@@ -110,7 +110,6 @@ namespace SchoolMedicalServer.Infrastructure.Services
                         }
 
                         var studentId = Guid.NewGuid();
-                        //var studentCode = row.Cell(1).GetString();
                         var studentCode = await studentRepository.GenerateStudentCodeAsync();
                         var fullName = row.Cell(1).GetString();
                         DateOnly? dayOfBirth = null;
@@ -224,7 +223,7 @@ namespace SchoolMedicalServer.Infrastructure.Services
                             UpdatedAt = DateTime.UtcNow
                         };
 
-                        await vacctionDetailsRepository.AddAsync(vaccineDetail);
+                        await vaccinationDetailsRepository.AddAsync(vaccineDetail);
                     }
                     await baseRepository.SaveChangesAsync();
                 }
@@ -232,158 +231,6 @@ namespace SchoolMedicalServer.Infrastructure.Services
             catch (Exception ex)
             {
                 throw new Exception($"Error uploading vaccination details: {ex.Message}", ex);
-            }
-        }
-
-        public async Task<byte[]> ExportStudentsExcelFileAsync()
-        {
-            try
-            {
-                var students = await studentRepository.GetAllAsync();
-
-                using var workbook = new XLWorkbook();
-                var worksheet = workbook.Worksheets.Add("Students");
-
-                worksheet.Cell(1, 1).Value = "StudentCode";
-                worksheet.Cell(1, 1).Value = "FullName";
-                worksheet.Cell(1, 2).Value = "DayOfBirth";
-                worksheet.Cell(1, 3).Value = "Gender";
-                worksheet.Cell(1, 4).Value = "Grade";
-                worksheet.Cell(1, 5).Value = "Address";
-                worksheet.Cell(1, 6).Value = "ParentPhoneNumber";
-                worksheet.Cell(1, 7).Value = "ParentEmailAddress";
-
-                int row = 2;
-                foreach (var student in students)
-                {
-                    worksheet.Cell(row, 1).Value = student.StudentCode;
-                    worksheet.Cell(row, 2).Value = student.FullName;
-                    worksheet.Cell(row, 3).Value = student.DayOfBirth.HasValue
-                        ? student.DayOfBirth.Value.ToString("yyyy-MM-dd")
-                        : "";
-                    worksheet.Cell(row, 4).Value = student.Gender;
-                    worksheet.Cell(row, 5).Value = student.Grade;
-                    worksheet.Cell(row, 6).Value = student.Address ?? "";
-                    worksheet.Cell(row, 7).Value = student.ParentPhoneNumber;
-                    worksheet.Cell(row, 8).Value = student.ParentEmailAddress;
-                    row++;
-                }
-
-                using var stream = new MemoryStream();
-                workbook.SaveAs(stream);
-                return stream.ToArray();
-            }
-            catch (Exception ex)
-            {
-                throw new Exception($"Error exporting students: {ex.Message}", ex);
-            }
-        }
-
-        public async Task<byte[]> ExportMedicalInventoriesExcelFileAsync()
-        {
-            try
-            {
-                var inventories = await medicalInventoryRepository.GetAllAsync();
-
-                using var workbook = new XLWorkbook();
-                var worksheet = workbook.Worksheets.Add("MedicalInventories");
-
-                worksheet.Cell(1, 1).Value = "ItemId";
-                worksheet.Cell(1, 2).Value = "ItemName";
-                worksheet.Cell(1, 3).Value = "Category";
-                worksheet.Cell(1, 4).Value = "Description";
-                worksheet.Cell(1, 5).Value = "QuantityInStock";
-                worksheet.Cell(1, 6).Value = "UnitOfMeasure";
-                worksheet.Cell(1, 7).Value = "MinimumStockLevel";
-                worksheet.Cell(1, 8).Value = "MaximumStockLevel";
-                worksheet.Cell(1, 9).Value = "LastImportDate";
-                worksheet.Cell(1, 10).Value = "LastExportDate";
-                worksheet.Cell(1, 11).Value = "ExpiryDate";
-                worksheet.Cell(1, 12).Value = "Status";
-
-                int row = 2;
-                foreach (var inventory in inventories)
-                {
-                    worksheet.Cell(row, 1).Value = inventory.ItemId.ToString();
-                    worksheet.Cell(row, 2).Value = inventory.ItemName;
-                    worksheet.Cell(row, 3).Value = inventory.Category;
-                    worksheet.Cell(row, 4).Value = inventory.Description;
-                    worksheet.Cell(row, 5).Value = inventory.QuantityInStock;
-                    worksheet.Cell(row, 6).Value = inventory.UnitOfMeasure;
-                    worksheet.Cell(row, 7).Value = inventory.MinimumStockLevel;
-                    worksheet.Cell(row, 8).Value = inventory.MaximumStockLevel;
-                    worksheet.Cell(row, 9).Value = inventory.LastImportDate.HasValue
-                        ? inventory.LastImportDate.Value.ToString("yyyy-MM-dd HH:mm:ss")
-                        : "";
-                    worksheet.Cell(row, 10).Value = inventory.LastExportDate.HasValue
-                        ? inventory.LastExportDate.Value.ToString("yyyy-MM-dd HH:mm:ss")
-                        : "";
-                    worksheet.Cell(row, 11).Value = inventory.ExpiryDate.HasValue
-                        ? inventory.ExpiryDate.Value.ToString("yyyy-MM-dd")
-                        : "";
-                    worksheet.Cell(row, 12).Value = inventory.Status ? "Available" : "Unavailable";
-                    row++;
-                }
-
-                using var stream = new MemoryStream();
-                workbook.SaveAs(stream);
-                return stream.ToArray();
-            }
-            catch (Exception ex)
-            {
-                throw new Exception($"Error exporting medical inventories: {ex.Message}", ex);
-            }
-        }
-
-        public async Task<byte[]> ExportVaccinationDetailsExcelFileAsync()
-        {
-            try
-            {
-                var vaccinationDetails = await vacctionDetailsRepository.GetAllAsync();
-
-                using var workbook = new XLWorkbook();
-                var worksheet = workbook.Worksheets.Add("VaccinationDetails");
-
-                worksheet.Cell(1, 1).Value = "VaccineId";
-                worksheet.Cell(1, 2).Value = "VaccineCode";
-                worksheet.Cell(1, 3).Value = "VaccineName";
-                worksheet.Cell(1, 4).Value = "Manufacturer";
-                worksheet.Cell(1, 5).Value = "VaccineType";
-                worksheet.Cell(1, 6).Value = "AgeRecommendation";
-                worksheet.Cell(1, 7).Value = "BatchNumber";
-                worksheet.Cell(1, 8).Value = "ExpirationDate";
-                worksheet.Cell(1, 9).Value = "ContraindicationNotes";
-                worksheet.Cell(1, 10).Value = "Description";
-                worksheet.Cell(1, 11).Value = "CreatedAt";
-                worksheet.Cell(1, 12).Value = "UpdatedAt";
-
-                int row = 2;
-                foreach (var detail in vaccinationDetails)
-                {
-                    worksheet.Cell(row, 1).Value = detail.VaccineId.ToString();
-                    worksheet.Cell(row, 2).Value = detail.VaccineCode ?? "";
-                    worksheet.Cell(row, 3).Value = detail.VaccineName ?? "";
-                    worksheet.Cell(row, 4).Value = detail.Manufacturer ?? "";
-                    worksheet.Cell(row, 5).Value = detail.VaccineType ?? "";
-                    worksheet.Cell(row, 6).Value = detail.AgeRecommendation ?? "";
-                    worksheet.Cell(row, 7).Value = detail.BatchNumber ?? "";
-                    worksheet.Cell(row, 8).Value = detail.ExpirationDate.HasValue
-                        ? detail.ExpirationDate.Value.ToString("yyyy-MM-dd")
-                        : "";
-                    worksheet.Cell(row, 9).Value = detail.ContraindicationNotes ?? "";
-                    worksheet.Cell(row, 10).Value = detail.Description ?? "";
-                    worksheet.Cell(row, 11).Value = detail.CreatedAt.ToString("yyyy-MM-dd HH:mm:ss");
-                    worksheet.Cell(row, 12).Value = detail.UpdatedAt.ToString("yyyy-MM-dd HH:mm:ss");
-                    row++;
-                }
-
-                using var stream = new MemoryStream();
-                workbook.SaveAs(stream);
-                return stream.ToArray();
-            }
-            catch (Exception ex)
-            {
-                throw new Exception($"Error exporting vaccination details: {ex.Message}", ex);
             }
         }
     }
